+++ CHANGES.rst
@@ -29,6 +29,7 @@ Changelog
 
 **Changed**
 
+- #1249 Render attachments in report in worksheets too
 - #1243 ID Server Suffix Support for Retested ARs
 - #1240 Support action-specific `workflow_action` requests with named adapters
 - #1215 Do not copy CaptureDate and Result in retest analyses when created